+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Ability to select media search type (movies vs. tv or both) in General -> Settings
+  - This controls the flow for the rest of the program and disables one or the other if not set to both
+
 ## [1.1.7] - 2026-08-14
 
 ### Changed